Added 'greedy' buffer concept
Please first note that buffered streams used to call `coroutine.yield()` when using read("*a") at the start of read and at every 2048 bytes.

# Explanation
There was a bug that happened on OpenComputers that would have Fuchas wait for any event in during the boot sequence. Through varied testing, it was found the `security.lateInit()` function semmed to be at fault. During boot, this function only called `loadUserPermissions()` which itself deserialize by reading (with `*a` parameter) on the file stream. And the problem was from the stream yielding even if we're not currently in a process, which meant yielding would wait for an event.
A first mitigation was just not yielding on the first 2048 bytes which also made more sense in other cases as it would increase performance. This solution wouldn't work if the user permissions file was larger than 2048 bytes, thus it was decided to add a new flag named `greedy`.

# Feature
When `greedy` is set to `true`, a buffered stream will never yield when reading it.
`security.lateInit` now uses a greedy stream.